remove login from controller to controllerUser
