+++ routes/index.js
@@ -1,7 +1,6 @@
 const router = require('express').Router()
 const routerTask = require('./routesTask')
 const routerUser = require('./routesUser')
-const Controller = require('./../controllers/controller')
 const ControllerUser = require('./../controllers/controllerUser')
 const { authentication } = require('./../middlewares/authentication')
 const { adminAuth } = require('./../middlewares/authentication')
@@ -10,8 +9,8 @@ router.get('/', (req, res) => {
     res.send("Hello World")
 })
 router.post('/users', ControllerUser.addUser)
-router.post('/login', Controller.login)
-router.post('/loginGoogle', Controller.loginGoogle)
+router.post('/login', ControllerUser.login)
+router.post('/loginGoogle', ControllerUser.loginGoogle)
 router.use(authentication)
 router.get('/:email/user', ControllerUser.getByEmail )
 router.use('/tasks', routerTask)
